Im currently changing the code structure and architecture to use modules
