+++ UtilsAndRepository/build.gradle
@@ -3,7 +3,6 @@ plugins {
     id 'kotlin-android'
     id 'kotlin-kapt'
     id 'kotlin-android-extensions'
-    id 'com.google.gms.google-services'
     id 'dagger.hilt.android.plugin'
 }
 
@@ -56,9 +55,11 @@ dependencies {
     kapt "androidx.hilt:hilt-compiler:1.0.0"
 
     //firebase
+/*
 
     implementation 'com.google.firebase:firebase-auth:20.0.4'
     implementation 'com.google.firebase:firebase-firestore:22.1.2'
+*/
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.4.2"
