+++ UtilsAndRepository/src/main/java/com/maricoolsapps/utilsandrepository/utils/ServerRepository.kt
@@ -1,8 +1,6 @@
 package com.maricoolsapps.utilsandrepository.utils
 
 import android.net.Uri
-import com.google.android.gms.tasks.Task
-import com.google.firebase.auth.AuthResult
 import com.maricoolsapps.serverdatabase.ServerUser
 import javax.inject.Inject
 
@@ -17,17 +15,17 @@ class ServerRepository
 
     val profilePhoto = user.getProfilePhoto()
 
-    fun changeProfilePhoto(newPhoto: Uri): Task<Void>? = user.changeProfilePhoto(newPhoto)
+    fun changeProfilePhoto(newPhoto: Uri) = user.changeProfilePhoto(newPhoto)
 
-    fun changeName(name: String): Task<Void>? = user.changeUsername(name)
+    fun changeName(name: String) = user.changeUsername(name)
 
-    fun reAuthenticate(oldPassword: String): Task<Void>?  = user.reAuthenticate(oldPassword)
+    fun reAuthenticate(oldPassword: String) = user.reAuthenticate(oldPassword)
 
-    fun changePassword(newPassword: String): Task<Void>? = user.changePasword(newPassword)
+    fun changePassword(newPassword: String)= user.changePasword(newPassword)
 
-    fun logInUser(email: String, password: String): Task<AuthResult> = user.signInUser(email, password)
+    fun logInUser(email: String, password: String) = user.signInUser(email, password)
 
-    fun createUser(email: String, password: String): Task<AuthResult>  = user.registerUser(email, password)
+    fun createUser(email: String, password: String)  = user.registerUser(email, password)
 
-    fun updateProfileName(name: String): Task<Void>? = user.updateProfileName(name)
+    fun updateProfileName(name: String) = user.updateProfileName(name)
 }
