+++ localDatabase/src/main/java/com/maricoolsapps/localdatabase/room/hilt_Modules.kt
@@ -1,0 +1,35 @@
+package com.maricoolsapps.localdatabase.room
+
+import android.content.Context
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.android.qualifiers.ApplicationContext
+import dagger.hilt.components.SingletonComponent
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.SupervisorJob
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+object hilt_Modules {
+
+    @Singleton
+    @Provides
+    fun provideMyRoomDatabase(@ApplicationContext context: Context): OnlineDatabase {
+        return OnlineDatabase.getDatabase(context)
+    }
+
+    @Singleton
+    @Provides
+    fun getApplicationScope(): CoroutineScope {
+        return CoroutineScope(SupervisorJob())
+    }
+
+    @Singleton
+    @Provides
+    fun provideDao(myRoomDatabase: OnlineDatabase): RoomDao {
+        return myRoomDatabase.dao()
+    }
+
+}
